refactor: cut and clean the code
