+++ libs/atqr/domain/src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/player/player.entity';
 export * from './lib/challenge/challenge.entity';
 export * from './lib/EmailService';
 export * from './lib/PaymentMethod';
+export * from './lib/player';
