--- libs/atqr/domain/src/lib/deadline-interface.ts
@@ -1,7 +1,0 @@
-import { Guid } from "@tokilabs/lang";
-import { Challenge } from "..";
-
-export interface DeadlineInterface {
-  id: Guid;
-  deadline: Date;
-}
